requirements.txt
sentence-transformers==5.2.0
transformers==4.57.4
torch==2.9.1
chromadb==1.4.0
langchain==0.1.0
langchain-text-splitters==0.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
reportlab>=4.0.0
requests>=2.31.0
streamlit>=1.28.0

# Module 5 dependencies
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
